support for multiple connect retries
This patch changes the way sinan connects to the server after it attempts to start. If the client is unable to connect to the server it will try to start it. It will then make multiple attempts to connect with a small interval between each one. This should allow for connecting as quickly as it can but still letting it start on slower boxes.