FIX 118: Show compiler options when building
The compiler options are not shown anywhere when building, so I added
a log message to show the options before each application is compiled.
This is needed because if using the compile_args option in _build.cfg,
there's no way to know if it actually works. It's also really useful
to know which flags are being used.